requirements.txt
streamlit
transformers
textsum
pdfminer.six